third_party/blink/web_tests/http/tests/images/feature-policy-legacy-formats-expected.txt

CONSOLE ERROR: Feature policy violation: legacy-image-formats is not allowed in this document.
